requirements.txt
python-dotenv>=1.0.1
openai>=1.30.1
google-generativeai>=0.5.4
groq>=0.6.0